Refactor telemetry
Move various members to `ApplicationTelemetry`.
