+++ src/API/Extensions/ILoggingBuilderExtensions.cs
@@ -24,9 +24,9 @@ public static class ILoggingBuilderExtensions
             options.IncludeFormattedMessage = true;
             options.IncludeScopes = true;
 
-            options.SetResourceBuilder(TelemetryExtensions.ResourceBuilder);
+            options.SetResourceBuilder(ApplicationTelemetry.ResourceBuilder);
 
-            if (TelemetryExtensions.IsOtlpCollectorConfigured())
+            if (ApplicationTelemetry.IsOtlpCollectorConfigured())
             {
                 options.AddOtlpExporter();
             }
