+++ src/API/Middleware/PyroscopeK6Middleware.cs
@@ -1,11 +1,6 @@
 ﻿// Copyright (c) Martin Costello, 2016. All rights reserved.
 // Licensed under the MIT license. See the LICENSE file in the project root for full license information.
 
-//// Based on https://github.com/grafana/pyroscope-go/blob/8fff2bccb5ed5611fdb09fdbd9a727367ab35f39/x/k6/baggage.go
-
-using OpenTelemetry;
-using Pyroscope;
-
 namespace MartinCostello.Api.Middleware;
 
 /// <summary>
@@ -22,53 +17,7 @@ internal sealed class PyroscopeK6Middleware(RequestDelegate next)
     /// <returns>
     /// A <see cref="Task"/> representing the actions performed by the middleware.
     /// </returns>
-    public async Task InvokeAsync(HttpContext context)
-    {
-        if (ExtractK6Baggage() is { Count: > 0 } baggage)
-        {
-            try
-            {
-                Profiler.Instance.ClearDynamicTags();
-
-                foreach ((string key, string value) in baggage)
-                {
-                    Profiler.Instance.SetDynamicTag(key, value);
-                }
-
-                await _next(context).ConfigureAwait(false);
-            }
-            finally
-            {
-                Profiler.Instance.ClearDynamicTags();
-            }
-        }
-        else
-        {
-            await _next(context).ConfigureAwait(false);
-        }
-    }
-
-    private static Dictionary<string, string>? ExtractK6Baggage()
-    {
-        if (Baggage.GetBaggage() is not { Count: > 0 } baggage)
-        {
-            return null;
-        }
-
-        Dictionary<string, string>? labels = null;
-
-        foreach ((string key, string? value) in baggage.Where((p) => p.Key.StartsWith("k6.", StringComparison.Ordinal)))
-        {
-            if (value is { Length: > 0 })
-            {
-                string label = key.Replace('.', '_');
-
-                // See https://github.com/grafana/jslib.k6.io/blob/80255ea7b239d3d4a8cd4e82192eef4ba27941a2/lib/http-instrumentation-pyroscope/1.0.1/index.js#L11
-                labels ??= new(3);
-                labels[label] = value;
-            }
-        }
-
-        return labels;
-    }
+    [System.Diagnostics.StackTraceHidden]
+    public Task InvokeAsync(HttpContext context) =>
+        ApplicationTelemetry.ProfileAsync((_next, context), (state) => state._next(state.context));
 }
